change users id

+++ server/app/models/User.scala
@@ -2,7 +2,7 @@ package models
 
 import play.api.libs.json._
 
-case class User(id: Long, name: String, password: String)
+case class User(id: Int, name: String, password: String)
 
 object User {
   implicit val userFormat = Json.format[User]
